requirements.txt
django==1.11.1
djangorestframework==3.6.3
django-webpack-loader